Include crate ts-rs dependencies when needed, move some views to account_management views
