+++ Cargo.toml
@@ -51,6 +51,7 @@ members = [
   "crates/utils",
   "crates/db_schema",
   "crates/db_schema_file",
+  "crates/db_views/account_management",
   "crates/db_views/api_misc",
   "crates/db_views/private_message",
   "crates/db_views/local_user",
@@ -119,6 +120,7 @@ lemmy_api_utils = { version = "=1.0.0-alpha.5", path = "./crates/api/api_utils" 
 lemmy_routes = { version = "=1.0.0-alpha.5", path = "./crates/routes" }
 lemmy_federate = { version = "=1.0.0-alpha.5", path = "./crates/federate" }
 lemmy_email = { version = "=1.0.0-alpha.5", path = "./crates/email" }
+lemmy_db_views_account_management = { version = "=1.0.0-alpha.5", path = "./crates/db_views/account_management" }
 lemmy_db_views_api_misc = { version = "=1.0.0-alpha.5", path = "./crates/db_views/api_misc" }
 lemmy_db_views_comment = { version = "=1.0.0-alpha.5", path = "./crates/db_views/comment" }
 lemmy_db_views_community = { version = "=1.0.0-alpha.5", path = "./crates/db_views/community" }
