+++ crates/api/api/Cargo.toml
@@ -19,6 +19,7 @@ doctest = false
 workspace = true
 
 [dependencies]
+lemmy_db_views_account_management = { workspace = true }
 lemmy_db_views_api_misc = { workspace = true, features = ["full"] }
 lemmy_db_views_comment = { workspace = true, features = ["full"] }
 lemmy_db_views_community = { workspace = true, features = ["full"] }
