+++ crates/api/api/src/local_user/change_password.rs
@@ -5,7 +5,7 @@ use actix_web::{
 use bcrypt::verify;
 use lemmy_api_utils::{claims::Claims, context::LemmyContext, utils::password_length_check};
 use lemmy_db_schema::source::{local_user::LocalUser, login_token::LoginToken};
-use lemmy_db_views_api_misc::{ChangePassword, LoginResponse};
+use lemmy_db_views_account_management::{ChangePassword, LoginResponse};
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
