+++ crates/api/api/src/local_user/change_password_after_reset.rs
@@ -5,7 +5,8 @@ use lemmy_db_schema::source::{
   login_token::LoginToken,
   password_reset_request::PasswordResetRequest,
 };
-use lemmy_db_views_api_misc::{PasswordChangeAfterReset, SuccessResponse};
+use lemmy_db_views_account_management::PasswordChangeAfterReset;
+use lemmy_db_views_api_misc::SuccessResponse;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
 
 pub async fn change_password_after_reset(
