+++ crates/api/api/src/local_user/generate_totp_secret.rs
@@ -6,7 +6,7 @@ use lemmy_db_schema::source::{
   local_user::{LocalUser, LocalUserUpdateForm},
   site::Site,
 };
-use lemmy_db_views_api_misc::GenerateTotpSecretResponse;
+use lemmy_db_views_account_management::GenerateTotpSecretResponse;
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
