+++ crates/api/api/src/local_user/get_captcha.rs
@@ -11,7 +11,7 @@ use actix_web::{
 use captcha::{generate, Difficulty};
 use lemmy_api_utils::context::LemmyContext;
 use lemmy_db_schema::source::captcha_answer::{CaptchaAnswer, CaptchaAnswerForm};
-use lemmy_db_views_api_misc::{CaptchaResponse, GetCaptchaResponse};
+use lemmy_db_views_account_management::{CaptchaResponse, GetCaptchaResponse};
 use lemmy_db_views_site::SiteView;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
