+++ crates/api/api/src/local_user/reset_password.rs
@@ -1,6 +1,7 @@
 use actix_web::web::{Data, Json};
 use lemmy_api_utils::{context::LemmyContext, utils::check_email_verified};
-use lemmy_db_views_api_misc::{PasswordReset, SuccessResponse};
+use lemmy_db_views_account_management::PasswordReset;
+use lemmy_db_views_api_misc::SuccessResponse;
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_db_views_site::SiteView;
 use lemmy_email::account::send_password_reset_email;
