+++ crates/api/api/src/local_user/update_totp.rs
@@ -2,7 +2,7 @@ use crate::check_totp_2fa_valid;
 use actix_web::web::{Data, Json};
 use lemmy_api_utils::context::LemmyContext;
 use lemmy_db_schema::source::local_user::{LocalUser, LocalUserUpdateForm};
-use lemmy_db_views_api_misc::{UpdateTotp, UpdateTotpResponse};
+use lemmy_db_views_account_management::{UpdateTotp, UpdateTotpResponse};
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_utils::error::LemmyResult;
