+++ crates/api/api/src/local_user/user_block_instance.rs
@@ -5,7 +5,8 @@ use lemmy_db_schema::{
   source::instance::{InstanceActions, InstanceBlockForm},
   traits::Blockable,
 };
-use lemmy_db_views_api_misc::{SuccessResponse, UserBlockInstanceParams};
+use lemmy_db_views_account_management::UserBlockInstanceParams;
+use lemmy_db_views_api_misc::SuccessResponse;
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
