+++ crates/api/api/src/local_user/verify_email.rs
@@ -4,7 +4,8 @@ use lemmy_db_schema::source::{
   email_verification::EmailVerification,
   local_user::{LocalUser, LocalUserUpdateForm},
 };
-use lemmy_db_views_api_misc::{SuccessResponse, VerifyEmail};
+use lemmy_db_views_account_management::VerifyEmail;
+use lemmy_db_views_api_misc::SuccessResponse;
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_db_views_site::SiteView;
 use lemmy_email::{account::send_email_verified_email, admin::send_new_applicant_email_to_admins};
