+++ crates/api/api_common/src/federation.rs
@@ -8,7 +8,8 @@ pub use lemmy_db_schema::{
   },
 };
 pub use lemmy_db_schema_file::enums::FederationMode;
-pub use lemmy_db_views_api_misc::{ResolveObject, UserBlockInstanceParams};
+pub use lemmy_db_views_account_management::UserBlockInstanceParams;
+pub use lemmy_db_views_api_misc::ResolveObject;
 pub use lemmy_db_views_readable_federation_state::ReadableFederationState;
 pub use lemmy_db_views_site::api::{
   FederatedInstances,
