+++ crates/api/api_crud/src/user/delete.rs
@@ -12,7 +12,8 @@ use lemmy_db_schema::source::{
   oauth_account::OAuthAccount,
   person::Person,
 };
-use lemmy_db_views_api_misc::{DeleteAccount, SuccessResponse};
+use lemmy_db_views_account_management::DeleteAccount;
+use lemmy_db_views_api_misc::SuccessResponse;
 use lemmy_db_views_local_user::LocalUserView;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
