+++ crates/api/api_crud/src/user/my_user.rs
@@ -10,7 +10,7 @@ use lemmy_db_schema::{
   },
   traits::Blockable,
 };
-use lemmy_db_views_api_misc::MyUserInfo;
+use lemmy_db_views_account_management::MyUserInfo;
 use lemmy_db_views_community_follower::CommunityFollowerView;
 use lemmy_db_views_community_moderator::CommunityModeratorView;
 use lemmy_db_views_local_user::LocalUserView;
