+++ crates/db_views/api_misc/Cargo.toml
@@ -17,23 +17,14 @@ workspace = true
 [features]
 full = [
   "lemmy_db_schema/full",
-  "lemmy_db_schema_file/full",
-  "lemmy_db_views_community_follower/full",
-  "lemmy_db_views_community_moderator/full",
-  "lemmy_db_views_local_user/full",
-  "lemmy_db_views_post/full",
   "lemmy_db_views_post/full",
   "extism",
   "extism-convert",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = ["dep:ts-rs", "lemmy_db_views_post/ts-rs"]
 
 [dependencies]
 lemmy_db_schema.workspace = true
-lemmy_db_schema_file.workspace = true
-lemmy_db_views_community_follower.workspace = true
-lemmy_db_views_community_moderator.workspace = true
-lemmy_db_views_local_user.workspace = true
 lemmy_db_views_post.workspace = true
 serde.workspace = true
 serde_with.workspace = true
