+++ crates/db_views/community_moderator/Cargo.toml
@@ -23,7 +23,7 @@ full = [
   "lemmy_db_schema/full",
   "lemmy_db_schema_file/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = ["dep:ts-rs", "lemmy_db_schema/ts-rs"]
 
 [dependencies]
 lemmy_db_schema = { workspace = true }
