+++ crates/db_views/inbox_combined/Cargo.toml
@@ -23,8 +23,13 @@ full = [
   "i-love-jesus",
   "lemmy_db_schema/full",
   "lemmy_db_schema_file/full",
+  "lemmy_db_views_private_message/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = [
+  "dep:ts-rs",
+  "lemmy_db_schema/ts-rs",
+  "lemmy_db_views_private_message/ts-rs",
+]
 
 [dependencies]
 lemmy_db_views_private_message = { workspace = true }
