+++ crates/db_views/local_user/Cargo.toml
@@ -25,7 +25,7 @@ full = [
   "lemmy_db_schema_file/full",
   "i-love-jesus",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = ["dep:ts-rs", "lemmy_db_schema/ts-rs"]
 
 [dependencies]
 lemmy_db_schema = { workspace = true }
