+++ crates/db_views/person/Cargo.toml
@@ -25,7 +25,11 @@ full = [
   "lemmy_db_schema_file/full",
   "lemmy_db_views_community_moderator/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = [
+  "dep:ts-rs",
+  "lemmy_db_schema/ts-rs",
+  "lemmy_db_views_community_moderator/ts-rs",
+]
 
 [dependencies]
 lemmy_db_schema = { workspace = true }
