+++ crates/db_views/person_content_combined/Cargo.toml
@@ -24,7 +24,13 @@ full = [
   "lemmy_db_schema/full",
   "lemmy_db_schema_file/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = [
+  "dep:ts-rs",
+  "lemmy_db_schema/ts-rs",
+  "lemmy_db_schema_file/ts-rs",
+  "lemmy_db_views_post/ts-rs",
+  "lemmy_db_views_comment/ts-rs",
+]
 
 [dependencies]
 lemmy_db_views_post = { workspace = true }
