+++ crates/db_views/post/Cargo.toml
@@ -26,7 +26,13 @@ full = [
   "lemmy_db_views_community/full",
   "lemmy_db_views_vote/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = [
+  "dep:ts-rs",
+  "lemmy_db_schema/ts-rs",
+  "lemmy_db_schema_file/ts-rs",
+  "lemmy_db_views_community/ts-rs",
+  "lemmy_db_views_vote/ts-rs",
+]
 
 [dependencies]
 lemmy_db_schema = { workspace = true }
