+++ crates/db_views/report_combined/Cargo.toml
@@ -24,7 +24,7 @@ full = [
   "lemmy_db_schema/full",
   "lemmy_db_schema_file/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = ["dep:ts-rs", "lemmy_db_schema/ts-rs", "lemmy_db_views_reports/ts-rs"]
 
 [dependencies]
 lemmy_db_views_reports = { workspace = true }
