+++ crates/db_views/search_combined/Cargo.toml
@@ -28,7 +28,15 @@ full = [
   "lemmy_db_views_community/full",
   "lemmy_db_views_person/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = [
+  "dep:ts-rs",
+  "lemmy_db_schema/ts-rs",
+  "lemmy_db_schema_file/ts-rs",
+  "lemmy_db_views_comment/ts-rs",
+  "lemmy_db_views_community/ts-rs",
+  "lemmy_db_views_person/ts-rs",
+  "lemmy_db_views_post/ts-rs",
+]
 
 [dependencies]
 lemmy_db_views_post = { workspace = true }
