+++ crates/db_views/site/Cargo.toml
@@ -26,7 +26,14 @@ full = [
   "lemmy_db_views_person/full",
   "lemmy_db_views_readable_federation_state/full",
 ]
-ts-rs = ["dep:ts-rs"]
+ts-rs = [
+  "dep:ts-rs",
+  "lemmy_db_schema/ts-rs",
+  "lemmy_db_schema_file/ts-rs",
+  "lemmy_db_views_api_misc/ts-rs",
+  "lemmy_db_views_person/ts-rs",
+  "lemmy_db_views_readable_federation_state/ts-rs",
+]
 
 [dependencies]
 lemmy_db_schema = { workspace = true }
